Adding XRLightEstimation to add light estimation capability
Light estimation is one of the key features of AR platforms (ARCore, ARKit,...),
and is essential for creating realistic AR contents.

The LightEstimation of the AR platforms are tightly coupled with the "frame",
because we can estimate the light of the AR scene with the environment
(camera preview ) of the scene.

It consists of the intensity and the color,
and the Application can apply them into the intensity and the color of the ambient light.
Color is optional because some platforms may support only the intensity.